rename data files
